fix: исправлена ошибка авторизации
